+++ docs/changelog.md
@@ -2,6 +2,17 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+# [0.9.0] - 2026-01-31
+## Added:
+- Processing of gse series bed files #143
+- Added metadata extraction from description if cell line or assay is not avaliable
+
+## Changed:
+- Improved reference genome predictor
+
+## Fixed:
+- Fix reference genome update and upload in bedbase db #144
+
 # [0.8.2] - 2025-09-22
 ## Fixed:
 - Downloading Refgenie chrom sizes (due to endpoint change)